fix: add types, styles, & snapshot for BasicFunctionality component
