+++ app/components/UI/BasicFunctionality/BasicFunctionality.styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+const styles = StyleSheet.create({
+  setting: {
+    marginVertical: 16,
+  },
+  heading: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    paddingBottom: 8,
+  },
+});
+
+export default styles;
