+++ app/components/UI/BasicFunctionality/BasicFunctionality.types.ts
@@ -1,0 +1,3 @@
+export interface BasicFunctionalityComponentProps {
+  handleSwitchToggle: () => void;
+}
